Updated the command line parser to allow the q heuristic to be used in the ascending trial.
Made the statistics that are print out be doubles instead of longs.

Fixed a bug in the filename of hrss trial that occurs on windows.

Added a gradle task for all vertices on n trial.

+++ Java/LabelAlgorithm/src/main/java/edu/csuci/trial/TrialStats.java
@@ -31,8 +31,8 @@ public class TrialStats {
 
     public void printStats(PrintStream output) {
         output.printf("Total Time: %d ms\n", time);
-        output.printf("Average Edges per Graph: %d Edges\n", edges/iterations);
-        output.printf("Average Labels per Graph: %d Labels\n", labels/iterations);
-        output.printf("Average Time per Graph: %d ms\n", time/iterations);
+        output.printf("Average Edges per Graph: %f Edges\n", (double) edges/iterations);
+        output.printf("Average Labels per Graph: %f Labels\n", (double) labels/iterations);
+        output.printf("Average Time per Graph: %f ms\n", (double) time/iterations);
    }
 }
